Initial commit
I only initialized this repo after having worked on the project for a
considerable while. In order to decrease whelm, I'm going to try to
perform piece-wise commits, but that means some things might break when
a commit happens, because it's dependent on something that doesn't exist
yet. I apologize for this.

In the beginning, I've only stored the Maven/git related files and the
WebauthnApplication entry point. However, pom.xml being as complicated
as it is... this is the reason.