Add WebAuthnProcessingFilter and configurer
The purpose of the WebAuthnProcessingFilter is to serve as a way for
Spring security to call the authentication providers with the correct
kinds of authentication tokens. The filter is put in the line of
processing filters, and acts as a UsernamePasswordAuthenticationFilter
except it can also choose to run the WebAuthn authentication provider's
authentication protocol if the correct credentials are in place.

The WebAuthnProcessingFilterConfigurer class is responsible for
configuring this filter. Currently it's as bare-bones an implementation
as it can be, but more configuration options can be added as necessary.

Importantly, I also need to retrieve a ServerProperty object based on
HttpServletRequest and HttpServletResponse objects, if the
spring-security-webauthn sample server implementation is to be believed.
I might need a ServerPropertyProvider class for this, so there is still
a lot of things to implement before this becomes truly usable.
The ServerProperty class is provided by the WebAuthn4J dependency.