Add /authenticate url for WebAuthn authentication
Created a separate url request pattern /authenticate for authenticating
with the WebAuthn API, which is where the user is redirected to if they
successfully log in using their username and password. This contains a
hidden form, which is submitted upon a call to
navigator.credentials.get() to the /login url, thereby allowing the
WebAuthProcessingFilter to trigger again and go through the second
branch (where the user's authenticator is validated).

I still need to implement authenticator registration
(navigator.credentials.create() + server side code) and actually
submit the form upon getting the credentials. After that, it should in
theory be possible to see the WebAuthn protocol in action in a
relatively minimalistic sample application including both the server
side and the client side code strictly necessary to make things work.

TODO: Add JavaScript files