Add/update static resources
This is part of a larger update/refactoring process. Some new things:
- Registration page added (register.html)
- Home page now links to either login page or registration page
- Common JS factored out into common.js (for now, only one function,
  loadChallenge())
- Titles on all pages updated

The static resources have changed, but the controls aren't entirely up
to date yet. I'll probably just add all of it in the next commit,
because I can't be bothered to sort out the dependencies of each file
and how to atomize my commits.

TODO: Figure out a way for the user to be redirected to the user page
after logging in, because for now it just constantly redirects to
/authenticate